+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Changelog
+
+## [1.0.0] - 2024-03-20
+
+### Added
+- Added initial holo loader files to package.
+- Initial README created